+++ test/oauth2-service.test.js
@@ -280,6 +280,33 @@ describe('OAuth 2 service', () => {
     });
   });
 
+  it('should allow customizing the token response through a beforeTokenSigning event', async () => {
+    service.once('beforeTokenSigning', (token, req) => {
+      expect(req).toBeInstanceOf(IncomingMessage);
+      /* eslint-disable no-param-reassign */
+      token.payload.custom_header = req.headers['custom-header'];
+      /* eslint-enable no-param-reassign */
+    });
+
+    const res = await tokenRequest(service.requestHandler)
+      .set('Custom-Header', 'custom-token-value')
+      .send({
+        grant_type: 'client_credentials',
+        scope: 'a-test-scope',
+      })
+      .expect(200);
+
+    const key = service.issuer.keys.get('test-rsa-key');
+
+    const decoded = jwt.verify(res.body.access_token, key.toPEM(false));
+
+    expect(decoded).toMatchObject({
+      iss: service.issuer.url,
+      scope: 'a-test-scope',
+      custom_header: 'custom-token-value',
+    });
+  });
+
   it('should expose the userinfo endpoint', async () => {
     const res = await request(service.requestHandler)
       .get('/userinfo')